Optimize compile time and rlib size.
Due to the way leptos works the typenames ends up being massive which in
turn creates massive rlib files for the linker to chew through.
In the case of the demo project the rlib file went from >500MB to around
91MB which also significantly speeds up incremental builds as linking
becomes the dominating part of the compilation time.

The way to alleviate this is by sprinkling in some strategic
`into_any()` usages at various points which basically type erases large
children names.