Minimise compile time and rlib size. (#48)
Due to the way leptos works the typenames ends up being massive which in
turn creates massive rlib files for the linker to chew through. In the
case of the demo project the rlib file went from >500MB to around 91MB
which also significantly speeds up incremental builds as linking becomes
the dominating part of the compilation time.

The way to alleviate this is by sprinkling in some strategic
`into_any()` usages at various points which basically type erases large
children names.

On my system this has taken incremental compile times of demo project
(using `trunk build`) to about 3.5s from 4.5s when reordering two
`Route` elements in `app.rs`.
However, the real gains are observed in my project where the rlib size
would increase to a gigantic 5GB which in turn would trigger this linker
bug and in turn make the project uncompilable:
https://github.com/rust-lang/rust/issues/130729.
With this fix the rlib sits at 1.3GB.